Appropriately set z-index (#70)
Indeed.com will block widget if containing `div` does not have a `z-index` value.

So set a `z-index` value for container such that it renders above any Indeed.com content and such that it does not block the tooltip's text (i.e. the container has a lower `z-index` than the tooltips).